+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Replace `log4j` with `logback` [#948] [#953]
+- Sort [`report`] violations by rule name within level [#955]
 
 ### Fixed
 - Fix custom [`report`] queries [#944]
@@ -293,6 +294,7 @@ First official release of ROBOT!
 [`template`]: http://robot.obolibrary.org/template
 [`validate`]: http://robot.obolibrary.org/validate
 
+[#955]: https://github.com/ontodev/robot/pull/955
 [#953]: https://github.com/ontodev/robot/pull/953
 [#951]: https://github.com/ontodev/robot/pull/951
 [#948]: https://github.com/ontodev/robot/pull/948