+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added new command: [`expand`] [#964]
 - Sort [`report`] violations by rule name within level [#955]
 
 ## [1.8.3] - 2021-12-16