+++ CHANGELOG.md
@@ -6,9 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Sort [`report`] violations by rule name within level [#955]
+
+## [1.8.3] - 2021-12-16
+
+### Changed
+- Replace `log4j` with `logback` [#948] [#953]
 
 ### Fixed
 - Fix custom [`report`] queries [#944]
+- Fix methods in `Report` object for ROBOT as a library [#951]
 
 ## [1.8.2] - 2021-12-02
 
@@ -246,7 +253,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 First official release of ROBOT!
 
-[Unreleased]: https://github.com/ontodev/robot/compare/v1.8.2...HEAD
+[Unreleased]: https://github.com/ontodev/robot/compare/v1.8.3...HEAD
+[1.8.3]: https://github.com/ontodev/robot/compare/v1.8.2...v1.8.3
 [1.8.2]: https://github.com/ontodev/robot/compare/v1.8.1...v1.8.2
 [1.8.1]: https://github.com/ontodev/robot/compare/v1.8.0...v1.8.1
 [1.8.0]: https://github.com/ontodev/robot/compare/v1.7.2...v1.8.0
@@ -286,6 +294,10 @@ First official release of ROBOT!
 [`template`]: http://robot.obolibrary.org/template
 [`validate`]: http://robot.obolibrary.org/validate
 
+[#955]: https://github.com/ontodev/robot/pull/955
+[#953]: https://github.com/ontodev/robot/pull/953
+[#951]: https://github.com/ontodev/robot/pull/951
+[#948]: https://github.com/ontodev/robot/pull/948
 [#944]: https://github.com/ontodev/robot/pull/944
 [#938]: https://github.com/ontodev/robot/pull/938
 [#929]: https://github.com/ontodev/robot/pull/929